refactor(tui): extract projects/slash-mirror mixins from tui_gateway/server.py (shard s5)
